fix(report): qualify hotspot function names with their container
`bca report markdown|html` rendered the bare AST name for each function,
so the ~20 per-language `impl Abc for <Lang>Code { fn compute(...) }`
methods in `src/metrics/abc.rs` all appeared as an indistinguishable
`compute` row in the published Pages report (they are distinct functions
with distinct line spans and complexity — not a double-count).

The hotspot walk now prefixes a nested function with its immediate
enclosing container, so the rows read `RustCode::compute`,
`KotlinCode::compute`, … — the same disambiguation `bca check` already
applies to offender lines.

- Extract the shared `space_segment` / `qualified_symbol` helpers out of
  `thresholds.rs` into a new `qualified_name` module. `bca check` keeps
  using the full `::`-joined chain (`qualified_symbol`) for precise
  offender identity; the report uses only the immediate container via
  `space_segment`, keeping each name O(1) so a pathologically deep AST
  cannot inflate report memory to O(depth²). For the shallow nesting real
  code produces, the two forms agree.
- The file-level (`Unit`) rows keep the file path as their name.
- `extract_summaries_inner`'s halstead.effort is suppressed with a
  rationale: it is the volume of the mechanical 30-field `CodeMetrics` ->
  `FunctionSummary` projection, not logic complexity.

Refreshed `.bca-baseline.toml` for the +1 line `mod` declaration and the
report walk's growth.
